Quick 'n dirty fix so that the code at least /compiles/ outside of VS

+++ source/ExternalDLL/ExternalDLL/Representation.cpp
@@ -99,10 +99,10 @@ bool Representation::calculateRepresentation(FeatureMap &features, double(&facia
 	double XN = getLength(X, N);
 	double XC = getLength(X, C);
 
-	double hl = std::abs(eyeLeft[0].x - eyeLeft[1].x);
-	double hy = std::abs(eyeRight[0].x - eyeRight[1].x);
-	double vl = std::abs(eyeLeft[0].y - eyeLeft[1].y);
-	double vy = std::abs(eyeRight[0].y - eyeRight[1].y);
+	double hl = std::abs((long)(eyeLeft[0].x - eyeLeft[1].x));
+	double hy = std::abs((long)(eyeRight[0].x - eyeRight[1].x));
+	double vl = std::abs((long)(eyeLeft[0].y - eyeLeft[1].y));
+	double vy = std::abs((long)(eyeRight[0].y - eyeRight[1].y));
 
 
 	facialParameters[0] = AB / OC; //Is relative, because divide
@@ -169,7 +169,7 @@ double Representation::getChinCountourArea(FeatureMap &features) {
 }
 
 double Representation::getVectorAngle(Point2D<double> one, Point2D<double> two) {
-	return std::acos((one.x * two.x) + (one.y * two.y));
+	return acos((one.x * two.x) + (one.y * two.y));
 }
 
 double Representation::getAngle(Point2D<double> one, Point2D<double> two, Point2D<double> orgin) {
@@ -179,7 +179,7 @@ double Representation::getAngle(Point2D<double> one, Point2D<double> two, Point2
 }
 
 double Representation::getVectorLenght(Point2D<double> vector) {
-	return std::sqrt(std::pow(vector.x, 2) + std::pow(vector.y, 2));
+	return sqrt(pow(vector.x, 2) + pow(vector.y, 2));
 }
 
 Point2D<double> Representation::getNormalizeVector(Point2D<double> vector) {
@@ -199,7 +199,7 @@ double Representation::getAreaOfTriangle(Point2D<double> A, Point2D<double> B, P
 	double s = (a + b + c) / 2.0;
 
 	//Area
-	return std::sqrt(s * (s - a) * (s - b) * (s - c));
+	return sqrt(s * (s - a) * (s - b) * (s - c));
 }
 
 Point2D<double> Representation::getCenter(Point2D<double> start, Point2D<double> end) {
@@ -207,5 +207,5 @@ Point2D<double> Representation::getCenter(Point2D<double> start, Point2D<double>
 }
 
 double Representation::getLength(Point2D<double> start, Point2D<double> end) {
-	return std::sqrt( std::pow((start.x - end.x), 2) + std::pow(start.y - end.y, 2) );
+	return sqrt(pow((start.x - end.x), 2) + pow(start.y - end.y, 2) );
 }
